refactor(direct-control): extract visibility read types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only visibility summary input/result type ownership and the reveal-map result type from packages/civ7-direct-control/src/index.ts into packages/civ7-direct-control/src/play/map/visibility.ts while keeping facade type re-exports in index.ts.

Preserves public type contracts and does not change visibility validation, bounded-grid semantics, approval-first disposable reveal behavior, reveal classification, source strings, runtime behavior, relationship-label policy, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- map-and-visibility.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.control.test.ts game.visibility.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, reveal behavior, relationship-label, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
