+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -404,6 +404,13 @@ runtime/direct-control claims.
         hotseat runtime proof, AI ingestion, CLI semantic projection,
         Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance
         pending.
+  - [x] 4.8.15 Extract visibility/reveal type owner while keeping public facade
+        type re-exports in `index.ts`, preserving visibility summary and reveal
+        map result public contracts while leaving setup, tactical, operation,
+        and ready result types, procedure schemas, operation/proof telemetry,
+        hotseat runtime proof, AI ingestion, CLI semantic projection,
+        Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance
+        pending.
 - [ ] 4.9 Import or explicitly cite oRPC architecture authority before any
       oRPC implementation; current support branch has no tracked
       `.agents/skills/civ7-orpc-control-architecture` or
