+++ openspec/changes/civ7-support-direct-control-modularization/workstream/direct-control-atom-corpus.md
@@ -172,6 +172,15 @@ AI ingestion, static profile shaping, procedure schemas, telemetry, hotseat
 runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task
 2.9.4 matrix-row acceptance.
 
+Visibility/reveal type update: `src/play/map/visibility.ts` now also owns
+visibility summary input/result types and the reveal-map result type while
+public facade type re-exports stay in `index.ts`. This is local package/type
+relocation proof only and does not change visibility validation, bounded-grid
+semantics, approval-first disposable reveal behavior, reveal classification,
+source strings, runtime behavior, relationship-label policy, telemetry, AI
+ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC
+procedure-core work, or Task 2.9.4 matrix-row acceptance.
+
 Bounded root inspection update: `src/runtime/root-inspection.ts` now owns
 `inspectCiv7Root` and the generated bounded root inspection command while the
 public facade export stays in `index.ts` and injects command execution,
